move record video function to data packet fragment
